separate Waker class to send_waker.h
